Change to abbreviation for "no date"
In the previous version of the style, the abreviation for "no dates" was "n.d." which is the english term. Although this is the term used in the provost Guidebook, it is incorrect to use this in french and should instead be "s.d." as is the case in other guidebooks of the french APA. Since many APA-french styles now point to Provost, I suggest that we update this here. If it is not possible to update it here (the provost guidebook does state that they use "n.d."), would it be possible to create another APA-french style to which others could point (such as UdeM and École de technologie supérieure)?